Redo MultiBuffer to handle multiple buffers, VPU now uses MultiBuffer
