+++ src/oakutils/nodes/buffer/_buffer.py
@@ -17,8 +17,6 @@ from typing import TYPE_CHECKING
 
 import numpy as np
 
-from ._multi_buffer import MultiBuffer
-
 if TYPE_CHECKING:
     import depthai as dai
     from typing_extensions import Self
@@ -54,7 +52,7 @@ class Buffer:
             output_streams = [output_stream]
         else:
             output_streams = output_stream
-        self._buffer = MultiBuffer(device, input_streams, output_streams)
+        self._buffer = _Buffer(device, input_streams, output_streams)
 
     def __call__(
         self: Self,
@@ -124,3 +122,106 @@ class Buffer:
         if len(data) == 1:
             return data[0]
         return data
+
+
+class _Buffer:
+    """Buffer for sending and receiving data from OAK-D."""
+
+    def __init__(
+        self: Self,
+        device: dai.DeviceBase,
+        input_streams: list[str],
+        output_streams: list[str],
+    ) -> None:
+        """
+        Create the buffer.
+
+        Parameters
+        ----------
+        device : dai.DeviceBase
+            The OAK-D device which the streams are built on.
+        input_streams : list[str]
+            A list of input stream names for the buffer to send data through,
+            when multiple inputs from host are needed.
+        output_streams : list[str]
+            The output stream names for the buffer to receive data from,
+            use multiple when want batches of data.
+
+        """
+        self._buffers: list[dai.Buffer] = []
+        self._input_queues: list[dai.DataInputQueue] = []
+        self._input_streams = input_streams
+        for stream in self._input_streams:
+            self._input_queues.append(device.getInputQueue(stream))  # type: ignore[attr-defined]
+            self._buffers.append(dai.Buffer())
+        self._output_queues: list[dai.DataOutputQueue] = []
+        self._output_streams = output_streams
+        for stream in self._output_streams:
+            self._output_queues.append(device.getOutputQueue(stream))  # type: ignore[attr-defined]
+
+    def __call__(
+        self: Self,
+        data: list[np.ndarray],
+    ) -> list[dai.ADatatype]:
+        """
+        Cycle data through the buffer.
+
+        Parameters
+        ----------
+        data : list[np.ndarray]
+            The list of data to cycle through the buffer.
+
+        Returns
+        -------
+        list[dai.ADatatype]
+            The data cycled through the buffer.
+
+        """
+        return self.cycle(data)
+
+    def cycle(
+        self: Self,
+        data: list[np.ndarray],
+    ) -> list[dai.ADatatype]:
+        """
+        Cycle data through the buffer.
+
+        Parameters
+        ----------
+        data : list[np.ndarray]
+            The list of data to cycle through the buffer.
+
+        Returns
+        -------
+        list[dai.ADatatype]
+            The data cycled through the buffer.
+
+        """
+        self.send(data)
+        return self.receive()
+
+    def send(self: Self, data: list[np.ndarray]) -> None:
+        """
+        Send data through the buffer.
+
+        Parameters
+        ----------
+        data : list[np.ndarray]
+            The list of data to send through the buffer.
+
+        """
+        for idx, d in enumerate(data):
+            self._buffers[idx].setData(d)
+            self._input_queues[idx].send(self._buffers[idx])
+
+    def receive(self: Self) -> list[dai.ADatatype]:
+        """
+        Receive data from the buffer.
+
+        Returns
+        -------
+        list[dai.ADatatype]
+            The data received from the buffer.
+
+        """
+        return [queue.get() for queue in self._output_queues]
