+++ src/oakutils/nodes/buffer/_multi_buffer.py
@@ -17,109 +17,99 @@ from typing import TYPE_CHECKING
 
 import depthai as dai
 
+from ._buffer import Buffer
+
 if TYPE_CHECKING:
     import numpy as np
     from typing_extensions import Self
 
 
 class MultiBuffer:
-    """Buffer for sending and receiving data from OAK-D."""
-
-    def __init__(
-        self: Self,
-        device: dai.DeviceBase,
-        input_streams: list[str],
-        output_streams: list[str],
-    ) -> None:
+    """Class for creating multiple Buffers for sending and receiving data from the OAK-D."""
+    def __init__(self: Self, device: dai.DeviceBase, input_streams: list[str | list[str]], output_streams: list[str | list[str]]) -> None:
         """
-        Create the buffer.
-
+        Create the multi buffer from streams.
+        
         Parameters
         ----------
         device : dai.DeviceBase
             The OAK-D device which the streams are built on.
-        input_streams : list[str]
-            A list of input stream names for the buffer to send data through,
-            when multiple inputs from host are needed.
-        output_streams : list[str]
-            The output stream names for the buffer to receive data from,
-            use multiple when want batches of data.
+        input_streams : list[str | list[str]]
+            The input stream names for the buffer to send data through.
+        output_streams : list[str | list[str]]
+            The output stream names for the buffer to receive data from.
 
         """
-        self._buffers: list[dai.Buffer] = []
-        self._input_queues: list[dai.DataInputQueue] = []
-        self._input_streams = input_streams
-        for stream in self._input_streams:
-            self._input_queues.append(device.getInputQueue(stream))  # type: ignore[attr-defined]
-            self._buffers.append(dai.Buffer())
-        self._output_queues: list[dai.DataOutputQueue] = []
-        self._output_streams = output_streams
-        for stream in self._output_streams:
-            self._output_queues.append(device.getOutputQueue(stream))  # type: ignore[attr-defined]
-
+        self._buffers: list[Buffer] = []
+        for input_stream, output_stream in zip(input_streams, output_streams):
+            self._buffers.append(Buffer(device, input_stream, output_stream))
+        
     def __call__(
         self: Self,
-        data: list[np.ndarray],
-    ) -> list[dai.ADatatype]:
+        data: list[np.ndarray | list[np.ndarray]],
+    ) -> list[dai.ADatatype | list[dai.ADatatype]]:
         """
-        Cycle data through the buffer.
-
+        Cycle data through the multi buffer.
+        
         Parameters
         ----------
-        data : list[np.ndarray]
-            The list of data to cycle through the buffer.
+        data : list[np.ndarray | list[np.ndarray]]
+            The data to cycle through the buffer.
 
         Returns
         -------
-        list[dai.ADatatype]
-            The data cycled through the buffer.
-
-        """
-        return self.cycle(data)
-
-    def cycle(
-        self: Self,
-        data: list[np.ndarray],
-    ) -> list[dai.ADatatype]:
-        """
-        Cycle data through the buffer.
-
-        Parameters
-        ----------
-        data : list[np.ndarray]
-            The list of data to cycle through the buffer.
-
-        Returns
-        -------
-        list[dai.ADatatype]
-            The data cycled through the buffer.
-
-        """
-        self.send(data)
-        return self.receive()
-
-    def send(self: Self, data: list[np.ndarray]) -> None:
-        """
-        Send data through the buffer.
-
-        Parameters
-        ----------
-        data : list[np.ndarray]
-            The list of data to send through the buffer.
-
-        """
-        for idx, d in enumerate(data):
-            self._buffers[idx].setData(d)
-            self._input_queues[idx].send(self._buffers[idx])
-
-    def receive(self: Self) -> list[dai.ADatatype]:
-        """
-        Receive data from the buffer.
-
-        Returns
-        -------
-        list[dai.ADatatype]
+        list[dai.ADatatype | list[dai.ADatatype]]
             The data received from the buffer.
 
         """
-        return [queue.get() for queue in self._output_queues]
+        return self.cycle(data)
+    
+    def cycle(
+        self: Self,
+        data: list[np.ndarray | list[np.ndarray]],
+    ) -> list[dai.ADatatype | list[dai.ADatatype]]:
+        """
+        Cycle data through the multi buffer.
+        
+        Parameters
+        ----------
+        data : list[np.ndarray | list[np.ndarray]]
+            The data to cycle through the buffer.
+
+        Returns
+        -------
+        list[dai.ADatatype | list[dai.ADatatype]]
+            The data received from the buffer.
+
+        """
+        return [buffer.cycle(data) for buffer in self._buffers]
+    
+    def send(
+        self: Self,
+        data: list[np.ndarray | list[np.ndarray]],
+    ) -> None:
+        """
+        Send data through the multi buffer.
+        
+        Parameters
+        ----------
+        data : list[np.ndarray | list[np.ndarray]]
+            The data to send through the buffer.
+
+        """
+        for buffer, data in zip(self._buffers, data):
+            buffer.send(data)
+
+    def receive(
+        self: Self,
+    ) -> list[dai.ADatatype | list[dai.ADatatype]]:
+        """
+        Receive data from the multi buffer.
+
+        Returns
+        -------
+        list[dai.ADatatype | list[dai.ADatatype]]
+            The data received from the buffer.
+
+        """
+        return [buffer.receive() for buffer in self._buffers]
